requirements.txt
SQLAlchemy
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
charset-normalizer
dataclasses-json
diskcache
exceptiongroup
frozenlist
greenlet
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-experimental
langsmith
llama-cpp-python
marshmallow
multidict
mypy-extensions
numpy
packaging
pydantic
pydantic-core
sniffio
regex
requests
tenacity
tiktoken
typing-extensions
typing-inspect
yarl